+++ safe_ice/core/__init__.py
@@ -5,11 +5,13 @@ from .ice_vmfnm import ICEvMFNM
 from .parameters import vMFNMParameters
 from .safe_ice import SafeICE
 from .safe_ice_optimized import OptimizedSafeICE
+from .subset_simulation import SubsetSimulation
 
 __all__ = [
     "AdaptiveSafeICE",
     "ICEvMFNM",
     "OptimizedSafeICE",
     "SafeICE",
+    "SubsetSimulation",
     "vMFNMParameters",
 ]